enable support for develop and pep517
