+++ MANIFEST.in
@@ -1,5 +1,7 @@
 include LICENSE
 include README*
+include MANIFEST.in
+include pyproject.toml
 include tox.ini
 include scripts/*
 include tests/*py
