+++ pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+# Further build requirements come from setup.py via the PEP 517 interface
+requires = [
+    "setuptools>=42",
+    "wheel",
+]
+build-backend = "setuptools.build_meta"
