+++ setup.cfg
@@ -1,4 +1,7 @@
 [egg_info]
 tag_build = .dev0
-#tag_svn_revision = yes
-#tag_date = yes
+
+[bdist_wheel]
+#universal = 1
+#python-tag = py3
+#plat-name = manylinux1_x86_64
