+++ tox.ini
@@ -25,7 +25,6 @@ deps =
 whitelist_externals =
     bash
 commands =
-    {envpython} setup.py build
-    {envpython} setup.py install
+    {envpython} -m pip install .
     bash -c "failed=0; for test in tests/__main__.py; do echo $test; \
              {envpython} $test || failed=1; done; exit $failed"
